HARVESTER: remove comment and unused code
