+++ detail/host/index.vue
@@ -3,7 +3,6 @@ import Tabbed from '@/components/Tabbed';
 import Tab from '@/components/Tabbed/Tab';
 import metricPoller from '@/mixins/metric-poller';
 import { METRIC, NODE, HCI } from '@/config/types';
-import { HOSTNAME } from '@/config/labels-annotations';
 import { allHash } from '@/utils/promise';
 import MaintenanceModal from '@/list/host/maintenanceModal';
 import CordonModal from '@/list/host/cordonModal';
@@ -33,25 +32,10 @@ export default {
   async fetch() {
     const hash = {
       nodes:        this.$store.dispatch('virtual/findAll', { type: NODE }),
-      vms:          this.$store.dispatch('virtual/findAll', { type: HCI.VM }),
       hostNetworks: this.$store.dispatch('virtual/findAll', { type: HCI.NODE_NETWORK }),
     };
 
     const res = await allHash(hash);
-    const instanceMap = {};
-
-    (this.$store.getters['virtual/all'](HCI.VMI) || []).forEach((vmi) => {
-      const vmiUID = vmi?.metadata?.ownerReferences?.[0]?.uid;
-
-      if (vmiUID) {
-        instanceMap[vmiUID] = vmi;
-      }
-    });
-
-    this.rows = res.vms.filter((row) => {
-      return instanceMap[row.metadata?.uid]?.status?.nodeName === this.value?.metadata?.labels?.[HOSTNAME];
-    });
-
     const hostNetowrkResource = res.hostNetworks.find( O => this.value.id === O.attachNodeName);
 
     this.loadMetrics();
@@ -65,7 +49,6 @@ export default {
     return {
       metrics:             null,
       mode:                'view',
-      rows:                [],
       hostNetowrkResource: null
     };
   },
@@ -96,7 +79,7 @@ export default {
         <Basic v-model="value" :metrics="metrics" :mode="mode" :host-netowrk-resource="hostNetowrkResource" />
       </Tab>
       <Tab name="instance" :label="t('harvester.host.tabs.instance')" :weight="2" class="bordered-table">
-        <Instance :rows="rows" />
+        <Instance :node="value" />
       </Tab>
     </Tabbed>
